parallel-pipeline/requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
requests>=2.31.0
psutil>=5.9.0
PyYAML>=6.0
redis>=5.0.0  # Optional, for Redis backend
Pillow>=10.0.0
